requirements.txt
torch==2.7.0 --index-url https://download.pytorch.org/whl/cu126
torchaudio==2.7.0 --index-url https://download.pytorch.org/whl/cu126
torchvision==0.22.0 --index-url https://download.pytorch.org/whl/cu126
opencv-python==4.11.0.86
tqdm==4.67.1
imageio==2.37.0
einops==0.8.1
scipy==1.15.2
matplotlib==3.10.0
ninja==1.11.1.3
numpy==1.26.4
pandas==2.2.3
huggingface_hub
wandb==0.19.8
timm==1.0.15
kornia==0.8.0
xformers==0.0.30
gradio_imageslider==0.0.20
gradio==4.29.0
OpenEXR
openai-clip